refactor: health관련 패키지 네이밍 및 파일 네이밍 변경
